grammar_scoring_engine/requirements.txt
flask
openai-whisper
language-tool-python
soundfile
numpy
torch
